requirements.txt

streamlit==1.37.0
google-generativeai==0.3.0
python-dotenv==1.0.0
gradio==5.20.0
flake8==6.1.0
pytest==7.4.0
pytest-mock==3.11.1
autopep8
isort
mypy
pylint
